Crate boot_info

This crate provides an abstraction over multiboot2 and UEFI boot information.

It allows the kernel’s initialisation to be the same, regardless of whether it was booted using BIOS or UEFI.

Modules

multiboot2

Structs

ElfSectionFlags

ELF section flags.

FramebufferInfo

Information about a framebuffer’s layout in memory.

ReservedMemoryRegion

Enums

FramebufferFormat

The format of the framebuffer, in graphical pixels or text-mode characters.

Traits

BootInformation

ElfSection

MemoryRegion

Module